Big bug fix. There were a few sinister compiler errors that luckily had simple solutions.
So, it turns out that for some reason, when I would create a string on the stack it would be created like so.
```
                          BAD
                           |
                           v
[ 'H' 'e' 'l' 'l' 'o' '!' '1' '0' ... ]
```

This was the original unexplained behaviour I mentioned in the readme, and this is fixed now. This means that a couple of the programs I had written myself no longer work BUT thats because those programs fit with this incorrect behaviour. Ive removed them from the `examples` directory.
This error was in the `ir::Value::string` function.
Additionally, there was a problem with the `cprint` builtin function that was faulty for this reason as well in the same file.

Another big change was the stack limit. After writing an `smpl` compiler in free, I found that the stack limit was reached very quickly. Write now I have a hard limit, but Im considering removing the limit entirely.

Additions
- the `getch` builtin function and the `Stdin` structure in `ir.rs`